+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.0.5
+
+- Stop storing provided points as they are not used by this package and can be stored by the
+  calling code if required.
+
 # 0.0.4
 
 - Add withinGeographicalRadius for calculating points within x kilometers from a given latitude and